% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSPP.R
\name{download_ospp}
\alias{download_ospp}
\title{Download data from open source psychometrics project}
\usage{
download_ospp(data_name)
}
\arguments{
\item{data_name:}{please specify the name of dataset}
}
\value{
return the data from open source psychometrics project
}
\description{
\code{download_ospp} load data from Open-Source Psychometrics Project(https://openpsychometrics.org/_rawdata/)
}
\examples{
#download_ospp("16PF")
}
