% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_openPsyData.R
\name{load_openPsyData}
\alias{load_openPsyData}
\title{Load data from open data for psychology}
\usage{
load_openPsyData(dataset_name, codebook = FALSE)
}
\arguments{
\item{data_name:}{please specify the name of dataset}

\item{codebook:}{If codebook set TRUE, load_openPsyData show the codebook of Open-Source Psychometrics Project}
}
\value{
return the data from open source psychometrics project
}
\description{
\code{load_openPsyData} load data from the repository of open data for psychology
}
\examples{
#data <- load_openPsyData("16PF", codebook = TRUE)
}
