% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSPP.R
\name{load_ospp}
\alias{load_ospp}
\title{Load data from open source psychometrics project}
\usage{
load_ospp(data_name, codebook = FALSE)
}
\arguments{
\item{data_name:}{please specify the name of dataset}

\item{codebook:}{If codebook set TRUE, load_ospp show the codebook of Open-Source Psychometrics Project}
}
\value{
return the data from open source psychometrics project
}
\description{
\code{load_ospp} load data from Open-Source Psychometrics Project(https://openpsychometrics.org/_rawdata/)
}
\examples{
#data <- load_ospp("16PF", codebook = TRUE)
}
